directory "uploaded_data" with some dummy pricelists and one example xml file

+++ rosetta_app/mc/model/table_items.class.php
@@ -119,7 +119,11 @@ class table_items
                                 <!-- Edit -->
                                 <div class="col-md-4 white col">
                                     <a href="edit_item.php?data_id=<?php echo $row['data_id']?>"><img src="lib/img/button_edit.png"/></a>
-                                    <a href="delete_item.php?data_id=<?php echo $row['data_id']?>"><img src="lib/img/button_delete.png"/></a>
+
+                                    <?php if($row["state"]=="active"){ ?>
+                                        <a href="delete_item.php?data_id=<?php echo $row['data_id']?>"><img src="lib/img/button_delete.png"/></a>
+                                    <?php } ?>
+
                                 </div>
                             </div>
